Frontend: workaround for iOS burger icon
iOS version utilizes the title property of Flyout page
as a 'hamburger icon'. Therefore, a png file for iOS
version has been added.

Workaround for https://github.com/dotnet/maui/issues/15258

The icon that is set for Flyout page is not
applicable to android version of the app
because of this bug:

https://github.com/dotnet/maui/issues/15211